+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbarItem.swift
@@ -3,7 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-03-18.
-//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//  Copyright © 2021-2023 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
@@ -93,9 +93,9 @@ struct AutocompleteToolbarItem_Previews: PreviewProvider {
     }
     
     static let previewSuggestions: [AutocompleteSuggestion] = [
-        StandardAutocompleteSuggestion(text: "Foo", isUnknown: true),
-        StandardAutocompleteSuggestion(text: "Bar", isAutocomplete: true),
-        StandardAutocompleteSuggestion(text: "", title: "Baz", subtitle: "Recommended")]
+        AutocompleteSuggestion(text: "Foo", isUnknown: true),
+        AutocompleteSuggestion(text: "Bar", isAutocomplete: true),
+        AutocompleteSuggestion(text: "", title: "Baz", subtitle: "Recommended")]
 }
 
 private extension View {