+++ Tests/KeyboardKitTests/Keyboard/KeyboardInputViewControllerTests.swift
@@ -234,7 +234,7 @@ class KeyboardInputViewControllerTests: XCTestCase {
     func testPerformingAutocompleteAbortsIfProxyIsReadingFullDocumentContext() {
         let vc = TestClass()
         setupMocksForAutocomplete(for: vc)
-        vc.autocompleteContext.suggestions = [StandardAutocompleteSuggestion(text: "")]
+        vc.autocompleteContext.suggestions = [AutocompleteSuggestion(text: "")]
         mockTextDocumentProxy.isReadingFullDocumentContext = true
         vc.performAutocomplete()
         mockTextDocumentProxy.isReadingFullDocumentContext = false
@@ -246,7 +246,7 @@ class KeyboardInputViewControllerTests: XCTestCase {
     func testPerformingAutocompleteAbortsIfTextProxyHasNoCurrentWord() {
         let vc = TestClass()
         setupMocksForAutocomplete(for: vc)
-        vc.autocompleteContext.suggestions = [StandardAutocompleteSuggestion(text: "")]
+        vc.autocompleteContext.suggestions = [AutocompleteSuggestion(text: "")]
         mockTextDocumentProxy.documentContextBeforeInput = nil
         mockTextDocumentProxy.documentContextAfterInput = nil
         XCTAssertNil(vc.autocompleteText)
@@ -259,7 +259,7 @@ class KeyboardInputViewControllerTests: XCTestCase {
     func testPerformingAutocompleteWritesResultToAutocompleteContext() {
         let vc = TestClass()
         setupMocksForAutocomplete(for: vc)
-        mockAutocompleteProvider.autocompleteSuggestionsResult = .success([StandardAutocompleteSuggestion(text: "")])
+        mockAutocompleteProvider.autocompleteSuggestionsResult = .success([AutocompleteSuggestion(text: "")])
         vc.performAutocomplete()
         eventually {
             XCTAssertEqual(vc.autocompleteContext.suggestions.count, 0)
@@ -267,7 +267,7 @@ class KeyboardInputViewControllerTests: XCTestCase {
     }
 
     func testResettingAutocompleteWritesResultToAutocompleteContext() {
-        vc.autocompleteContext.suggestions = [StandardAutocompleteSuggestion(text: "")]
+        vc.autocompleteContext.suggestions = [AutocompleteSuggestion(text: "")]
         vc.resetAutocomplete()
         XCTAssertEqual(vc.autocompleteContext.suggestions.count, 0)
     }